+++ compas_python_utils/detailed_evolution_plotter/plot_detailed_evolution.py
@@ -44,11 +44,13 @@ def run_main_plotter(data_path, outdir='.', show=True, use_latex=True):
     printEvolutionaryHistory(events=events)
 
     ### Produce the two plots
-    makeDetailedPlots(Data, events, outdir=outdir, use_latex=use_latex)
-    plotVanDenHeuvel(events=events, outdir=outdir, use_latex=use_latex)
+    detailed_fig = makeDetailedPlots(Data, events, outdir=outdir, use_latex=use_latex)
+    vdh_fig, vdh_events = plotVanDenHeuvel(events=events, outdir=outdir, use_latex=use_latex)
+
     if show:
         plt.show()
 
+    return detailed_fig, vdh_fig, vdh_events
 
 def set_font_params(use_latex=True):
     use_latex = use_latex and (shutil.which("latex") is not None)
@@ -127,7 +129,10 @@ def makeDetailedPlots(Data=None, events=None, outdir='.', show=True, use_latex=T
     fig.suptitle('Detailed evolution for seed = {}'.format(Data['SEED'][()][0]), fontsize=18)
     fig.tight_layout(h_pad=1, w_pad=1, rect=(0.08, 0.08, .98, .98), pad=0.)  # (left, bottom, right, top)
 
-    safe_save_figure(fig, f'{outdir}/detailedEvolutionPlot.png', bbox_inches='tight', pad_inches=0, format='png')
+    if outdir is not None:
+        safe_save_figure(fig, f'{outdir}/detailedEvolutionPlot.png', bbox_inches='tight', pad_inches=0, format='png')
+    
+    return fig
 
 
 ######## Plotting functions
@@ -353,9 +358,11 @@ def plotVanDenHeuvel(events=None, outdir='.', use_latex=True):
         axs[ii].annotate(chr(ord('@') + 1 + ii), xy=(-0.15, 0.8), xycoords='axes fraction', fontsize=8,
                          fontweight='bold')
 
-    file_path = os.path.join(outdir, 'vanDenHeuvelPlot.eps')
-    safe_save_figure(fig, file_path, bbox_inches='tight', pad_inches=0, format='eps')
-    return fig
+    if outdir is not None:
+        file_path = os.path.join(outdir, 'vanDenHeuvelPlot.eps')
+        safe_save_figure(fig, file_path, bbox_inches='tight', pad_inches=0, format='eps')
+
+    return fig, events
 
 
 ### Helper functions
@@ -450,7 +457,7 @@ class Event(object):
 
         self.eventImage = None
         self.endState = None  # sets the endstate - only relevant if eventClass=='End'
-        self.eventString = self.getEventDetails(use_latex=use_latex, **kwargs)
+        self.eventString, self.image_num, self.rotate_image = self.getEventDetails(use_latex=use_latex, **kwargs)
 
     def getEventDetails(self, use_latex=True, **kwargs):
         """
@@ -487,14 +494,16 @@ class Event(object):
                     image_num = 44
             elif mtValue == 3:
                 eventString = r'Common envelope initiated by 1'
-                if (self.stype1 < 13) & (self.stype2 < 13):
+                if self.stype2 < 13:
                     image_num = 28
                 else:
                     image_num = 49
+                    rotate_image = True
             elif mtValue == 4:
                 eventString = r'Common envelope initiated by 2'
-                if (self.stype1 < 13) & (self.stype2 < 13):
+                if self.stype1 < 13:
                     image_num = 28
+                    rotate_image = True
                 else:
                     image_num = 49
             elif mtValue == 5:
@@ -535,8 +544,12 @@ class Event(object):
             else:
                 if compType < 13:
                     image_num = 13  # 13 for normal companion
+                elif compType == 13:
+                    image_num = 15  # 15 for NS companion
                 else:
-                    image_num = 15  # 15 for CO companion
+                    image_num = 17  # 17 for BH companion
+            if whichStar == 2:
+                rotate_image = True
 
         elif eventClass == 'Stype':
             whichStar = kwargs['whichStar']
@@ -566,26 +579,28 @@ class Event(object):
                 T0 = a ** 4 / 4 / beta
                 Tdelay = T0 * (1 - e ** 2) ** (7 / 2) * (
                         1 + 0.31 * e ** 10 + 0.27 * e ** 20 + 0.2 * e ** 1000) / 3.15e7 / 1e6
-                eventString = r'Double compact object ({}+{}) merging in {:.2e} Myr'.format(self.stypeName1,
-                                                                                            self.stypeName2, Tdelay)
+                eventString = r'Double compact object ({}+{}) merging in {:.2e} Myr'.format(self.stypeName1, self.stypeName2, Tdelay)
+                self.time=self.time+Tdelay
 
-                if (stype1 == 13) & (stype2 == 13):
+                if (stype1 == 13) and (stype2 == 13):
                     image_num = 55
-                elif (stype1 == 14) & (stype2 == 14):
+                elif (stype1 == 14) and (stype2 == 14):
                     image_num = 51
                 else:
                     image_num = 53
+                    if (stype1 == 14) and (stype2 == 13):
+                        rotate_image = True
 
             elif state == "Unbound":
                 eventString = r'Unbound: {}+{}'.format(self.stypeName1, self.stypeName2)
-                if (stype1 == 13) & (stype2 < 13):
+                if (stype1 == 13) and (stype2 < 13):
                     image_num = 19
-                elif (stype1 < 13) & (stype2 == 13):
+                elif (stype1 < 13) and (stype2 == 13):
                     image_num = 19
                     rotate_image = True
-                elif (stype1 == 14) & (stype2 < 13):
+                elif (stype1 == 14) and (stype2 < 13):
                     image_num = 20
-                elif (stype1 < 13) & (stype2 == 14):
+                elif (stype1 < 13) and (stype2 == 14):
                     image_num = 20
                     rotate_image = True
                 else:
@@ -601,7 +616,7 @@ class Event(object):
         if image_num != None:
             self.eventImage = self.getEventImage(image_num, rotate_image)
 
-        return eventString
+        return eventString, image_num, rotate_image
 
     def getEventImage(self, image_num, rotate_image):
         """
@@ -710,7 +725,7 @@ def printEvolutionaryHistory(Data=None, events=None):
 
     for event in events:
         ii = event.index
-        printFormattedEvolutionLine(Data['Time'][ii], event.eventString.replace('$', ''),
+        printFormattedEvolutionLine(event.time, event.eventString.replace('$', ''),
                                     Data['Mass(1)'][ii], Data['Stellar_Type(1)'][ii],
                                     Data['Mass(2)'][ii], Data['Stellar_Type(2)'][ii],
                                     Data['SemiMajorAxis'][ii], Data['Eccentricity'][ii])