+++ src/HG.cpp
@@ -1203,9 +1203,12 @@ STELLAR_TYPE HG::EvolveToNextPhase() {
  *
  */
 void HG::UpdateInitialMass() {
-    // only update mass0 if the current mass would yield a core mass larger than or equal to the current core mass
+    // only update mass0 on mass loss if the current mass would yield a core mass larger than or equal to the current core mass
     // i.e., no unphysical core mass decrease would ensue
-    if (utils::Compare(m_CoreMass, HG::CalculateCoreMassOnPhaseIgnoringPreviousCoreMass(m_Mass, m_Age)) <= 0) {
+    // (we do not update mass0 on mass gain on the HG -- there is no instruction for doing so in Hurley; adding this
+    // check also avoid difficulties for the BRCEK rejuvenation prescription, when mass0 may be set to enforce a core mass
+    // that is lower than would be expected for the current mass value according to the Hurley prescription)
+    if (utils::Compare(m_Mass0, m_Mass) > 0 && utils::Compare(m_CoreMass, HG::CalculateCoreMassOnPhaseIgnoringPreviousCoreMass(m_Mass, m_Age)) <= 0) {
         m_Mass0 = m_Mass;
     }
 }